pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Slyfest <mellowstripe@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytesseract = "^0.3.7"
opencv-python = "^4.5.1"
flashtext = "^2.7"
python-telegram-bot = "^13.2"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipykernel = "^5.4.3"
requests = "^2.25.1"
bs4 = "^0.0.1"
pandas = "^1.2.1"
prettyprint = "^0.1.5"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
